feat(plugin): add integration hubstaff UI package
- Added new @gauzy/plugin-integration-hubstaff-ui package for integrating Hubstaff UI components
