--- apps/gauzy/src/app/pages/hubstaff/components/hubstaff-authorize/hubstaff-authorize.component.scss
@@ -1,1 +1,0 @@
-@import '../../../upwork/components/upwork-authorize/upwork-authorize.component.scss';
