+++ packages/plugins/integration-hubstaff-ui/src/lib/components/hubstaff/hubstaff.component.scss
