+++ packages/plugins/integration-hubstaff-ui/src/lib/components/settings-dialog/settings-dialog.component.scss
