+++ packages/plugins/integration-hubstaff-ui/src/lib/integration-hubstaff-routing.module.ts
@@ -1,0 +1,31 @@
+import { NgModule } from '@angular/core';
+import { RouterModule } from '@angular/router';
+import { HubstaffAuthorizeComponent } from './components/hubstaff-authorize/hubstaff-authorize.component';
+import { HubstaffComponent } from './components/hubstaff/hubstaff.component';
+
+@NgModule({
+	imports: [
+		RouterModule.forChild([
+			{
+				path: '',
+				component: HubstaffAuthorizeComponent,
+				data: {
+					state: true
+				}
+			},
+			{
+				path: 'regenerate',
+				component: HubstaffAuthorizeComponent,
+				data: {
+					state: false
+				}
+			},
+			{
+				path: ':id',
+				component: HubstaffComponent
+			}
+		])
+	],
+	exports: [RouterModule]
+})
+export class IntegrationHubstaffRoutingModule {}
