+++ packages/plugins/integration-hubstaff-ui/tsconfig.json
@@ -10,7 +10,9 @@
 		"noFallthroughCasesInSwitch": false,
 		"baseUrl": ".",
 		"paths": {
-			"@gauzy/ui-core/*": ["./../../../dist/packages/ui-core/*/index.d.ts"]
+			"@gauzy/ui-core/*": ["./../../../dist/packages/ui-core/*/index.d.ts"],
+			"@ngx-translate/*": ["./node_modules/@ngx-translate/*"],
+			"ngx-permissions": ["./node_modules/ngx-permissions"]
 		}
 	},
 	"files": [],
